UPSTREAM: device/pciexp_device: Set values numerically instead of as bitmask
As noted on linux-pci, we have a weird way to handling "value" and
"scale" fields that are supposed to contain numerical values: we encode
them as a bitfield.
Instead define the two fields (offset and mask) and use numbers.

Another issue, not fixed in this CL, is that we write hard-coded values
while these fields really need to contain the max() of acceptable delays
of the downstream devices. That way the controller can decide whether or
not to enter a deeper power management state. It's noted as a TODO.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie4f692b44afc6198121e5097b67015022705c227
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9adcbfe486cb37ecbf38167ac13a502326f5863c
Original-Change-Id: I895b9fe2ee438d3958c2d787e70a84d73eaa49d2
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Found-by: Bjorn Helgaas <bhelgaas@google.com>
Original-Reviewed-on: https://review.coreboot.org/22740
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/819004
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>