UPSTREAM: soc/intel/apollolake: Add SMI and SCI support for ESPI
This patch adds the SMI bits for SMI_EN, SMI_STS and GPE
register in pm.h. The southbridge handler for espi smi is
also added. In gpe.h we add GPE0A_ESPI_SCI_STS which is
bit 20 in GPE register and enables the setting of the
ESPI_SCI STS bit to generate a wake event and/or an SCI/SMI.

TEST=  Boot to OS.

Change-Id: If61214fdab6c3b10d7544e29c4fbcf8215089c3a
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 41cfd5ba69783c33883d4515ffa15ac279afb831
Original-Change-Id: I2b8372ffbe0949ddd4aa83bdd7c0a01ade3ed40e
Original-Signed-off-by: Shaunak Saha <shaunak.saha@intel.com>
Original-Reviewed-on: https://review.coreboot.org/22758
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>