UPSTREAM: soc/intel/skylake: Add integrated LAN config parameters
Add parameters to configure the integrated LAN via FSP.  Since
this takes over a PCI CLKREQ# pin it needs to know which pin
it should use, and there are additional parameters for LTR and
a "K1 power save" feature.

This was tested on a KBL-R board with integrated LAN, verifying
that the device is functional under Linux with the e1000e driver.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia1a34a6430ec76b0b5eaf69b34fb6a59466d3d3a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 14485efbb33e417330df59fcc501b336e14ef55c
Original-Change-Id: Idb200cec90a3c0d4d9c914bae9983a3bcdafcd06
Original-Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/22856
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/829785
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>