UPSTREAM: util/gitconfig: Make gitconfig a bash script
The gitconfig target has a few bashisms and would fail
silently on systems that use a POSIX standard sh (like Ubuntu dash).

Remove the code from the makefile and put it in a bash script that
is called by the gitconfig target.

BUG=none
BRANCH=none
TEST=none

Change-Id: I38d7072265dbf8904d5091ddf4175e7e4f6896f2
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 9ab8ae6a55681874ec194bc144077ded5363a317
Original-Change-Id: I3bc8cf688a3ad211b57c8ca0e6b1e86c82dc6a37
Original-Signed-off-by: Marc Jones <marcj303@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/22857
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/831646
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>