UPSTREAM: amd/stoneyridge: Force PSP command reg settings in bootblock
A subsequent patch to the PSP library will rely on the device already
having its PCI command register set to allow memory decoding and
mastering enabled.

Program the command register ahead of loading the SMU FW1 blob in
bootblock.  When the device has not been set up (e.g. when SMU FW
is not selectable), AGESA sets up the device.  As a result, a
similar change is not required before sending the DRAM ready
command.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic3cb46441c214b5d14f31df397a67a0343938d48
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 854d4dd9e2bd803a81ae0c90c26143a10f25abc4
Original-Change-Id: Id586106751286c4767b5c16ed7e1604523635492
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/22876
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>