UPSTREAM: google/gru: Adjust to incorrect strapping resistors on Kevin
It seems that RAM code 0 has been strapped with an incorrect resistor on
Kevin. The resulting voltage divide still puts it well within the ADC
value bucket reserved for that slot, but a little closer to the edge
than necessary. While this doesn't seem to cause any immediate problems
on its own, it still doesn't hurt to fix it (if only for the
documentation value).

On other boards (at least on my Scarlet) the strapping seems to be
correct.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id2cbfb572f27217539523b5fa43857cc85a2b328
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 621b022c3b0a0aaa93fa45e889472ded5a2a4557
Original-Change-Id: Ic5199834fbeaf734e725ff45b04f45eefe149855
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/22891
Original-Reviewed-by: David Schneider <dnschneid@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>