UPSTREAM: sb/intel/i82801ix: Don't select HAVE_INTEL_FIRMWARE on Q35-QEMU
The Qemu q35 target doesn't support or needs Intel Firmware blobs so
it doesn't make sense to select that option on this hardware.

The result of this change will be that when changing the ROM chip
size, CBFS_SIZE will automatically fill the whole flash which is
desirable in this case.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib65d1a7d780832ffe3d846ae3bd83bc5116d684f
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 5495ed2a4397bbd106f3df59a252b1cf0227d1a3
Original-Change-Id: I89b0c2a7b3e9c163ce4b4eb5b38ab5fa70ba3cfa
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/23090
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/852943
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>