UPSTREAM: soc/intel/common: Add Intel HDA common block driver
There is common HDA code in soc/intel/common that provides generic
HDA support functions, but it does not provide a driver.

This change adds a common block driver for HDA that provides a
ramstage driver for SOCs that need to initialize an HDA codec.

This was tested on a board with an HDA codec to ensure that it
properly detected it and ran the codec init steps.

BUG=none
BRANCH=none
TEST=none

Change-Id: I383f76dd41dab00a78b00d11c5109193a0e13888
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 93142a452e4c0cb35ea04394ded171b1cf58ff72
Original-Change-Id: I41b4c54d3c81e1f09810cfaf934ffacafca1cf38
Original-Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/23187
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/865531
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>