UPSTREAM: soc/amd/common/pi: Fix issue in AGESA heap allocator
The heap allocator would try to split a buffer node that
was too small for another node. In the failing case, the buffer
node was 0x140 bytes and the requested size was 0x133 bytes.
The logic would check that there was room for the header and
buffer and try to split the buffer node. The buffer node header
is 0xC bytes, so 0x13F bytes are need. The problem is that it didn't
leave room for another node header and a little space for a buffer.

BUG=b:71764350
TEST= Boot grunt.
BRANCH=none

Change-Id: I7d6deeff44560e78ea54314beaa51a78a355d375
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: ca966f9a2df171848e57f4deac0dda6ce24cc938
Original-Change-Id: Iece5e12d5787415a335bb953985331a5dc312152
Original-Signed-off-by: Marc Jones <marcj303@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/23211
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Reviewed-by: Daniel Kurtz <djkurtz@google.com>
Reviewed-on: https://chromium-review.googlesource.com/865528
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>