UPSTREAM: util/crossgcc: Output apt-get commands on debian
In the buildgcc script, there is a check that the tools required are
installed. When a tool is missing, a message is output suggesting an
installation method, e.g. `sudo apt-get install foo` on debian-based
systems.

When run on a true, vanilla debian system, the error message provides
only a generic hint because the `please_install()` function fails to
detect the OS kind. Detection is based on definition of `ID_LIKE` in
`/etc/os-release` yet such systems only define `ID` to `debian`.

This commit closes the detection gap. Tested on debian 9 (stretch).

BUG=none
BRANCH=none
TEST=none

Change-Id: I43bfcaf21e0acb5fa55253632ea9c99713a77088
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 72d1089fac4d68207c8360ab7b4663b41b16c44e
Original-Change-Id: I3c867837e9157bee13010bd0a005028c369ce55f
Original-Signed-off-by: Alex Thiessen <alex.thiessen.de+coreboot@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/23231
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/868769
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>