UPSTREAM: util/release: Improve git worktree checks
The bash script `genrelnotes` checks for `.git` to be present to
determine whether the current directory is the top directory of a git
worktree. This check is rather weak and doesn't handle many edge cases
like that of a broken gitfile.

Add a proper `git rev-parse` call to check the condition.

BUG=none
BRANCH=none
TEST=none

Change-Id: I8a4dd935b801b484575059644eb394356551c357
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: da384dd7a4eea2fc4153c593026489cee5585895
Original-Change-Id: I32b06ca982d55fd8e88e55651b6bc53014905823
Original-Signed-off-by: Alex Thiessen <alex.thiessen.de+coreboot@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/23252
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/871004
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>