UPSTREAM: util/genbuild_h: Add gitfile support
`genbuild_h.sh` checks whether it is running from a coreboot's own git
worktree to decide whether to use git as the time source. This check
fails when `${top}/.git` is a gitfile, e.g. when coreboot is a
submodule.

Add a proper `git rev-parse` call to check the condition, remove `$top`
which is not used anymore.

BUG=none
BRANCH=none
TEST=none

Change-Id: I046b6a1d3dfbf406bc179b7ec067ea070e9fce40
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1758fd2a32408b9206c41b8e71300494ee53f886
Original-Change-Id: I8bb13d607a01f4f28fa8b165769e0a1f702da362
Original-Signed-off-by: Alex Thiessen <alex.thiessen.de+coreboot@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/23256
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/881004
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>