UPSTREAM: google/scarlet: Add initialization sequence for Innolux P097PFG panel
Innolux didn't deliver a working init sequence yet for devices without
OTP programming. The sequence in this change has been derived from a
register dump of a mostly working panel with OTP. It is not meant to
be final, but to make devices with unprogrammed OTP work, while Innolux
is figuring out a proper sequence. There is a known issue with an
artifact line in the lower third of the display.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib9a4ed8e84cd3fc44bd9415cae870db726d27fa9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3c0d7cfb947d12bdadc5d1898d23d29c93e0ec03
Original-Change-Id: I7096506208e4cb29c5f31a7ac502231a6c23ac92
Original-Signed-off-by: Lin Huang <hl@rock-chips.com>
Original-Reviewed-on: https://review.coreboot.org/23311
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/892065
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>