UPSTREAM: soc/amd/stoneyridge: use new host controller programming
The SPI controller on stoneyridge apparently has a large fifo
and an alternate method for programming the controller. The
fifo is directly accessible as well as the rx and tx pointer
in addition to the execute bit. Remove the unneeded #defines
and program the host controller with the above changes in mind.

In addition, add debug hooks to the driver so one can dump the
state of the controller when in operation.

The time it took to read 4KiB of flash in the elog driver went
from 20593 microseconds to 5693 microseconds on cdx03/kahlee.

BUG=b:65485690

Change-Id: I8696ca0a83a1c9b7a320d1194d7091df00450967
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0d2d77a597542354a1fe64cad908f6a9e0d59591
Original-Change-Id: Ie7ea9d18cef5511686700ad9b2b9fdfeb6d5685b
Original-Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/23493
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Justin TerAvest <teravest@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/892068
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>