UPSTREAM: lib/stage_cache: Add save/get raw storage
Leverage the stage_cache mechanism to store a non-specific type
of data.  This is not interesting when the location for the cache
is in cbmem.  However it will be more useful when an external
location is used, e.g. when the cache is in TSEG, locked from user
modification.

BUG=none
BRANCH=none
TEST=none

Change-Id: I43896df41017a5abae58bc14c39424c20a4355f3
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 8d6e0e0a72ed81d44ba61add0c2aab55bb217412
Original-Change-Id: Iaf0b25ebe14c176bbd24fc8942f902f627ca8e6f
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/23518
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/907391
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>