UPSTREAM: soc/amd/stoneyridge/acpi/sb_pci0_fch.asl: Fix instability
A file that has several methods cannot be included inside a method. It has
to be included inside a scope, but not inside a method or it'll cause
problems (instability).

There is an ugly construction in method _INI. It's needed because if AmdImc
is not included then the call to ITZE would break the build.

BUG=b:62200858
TEST=Build kahlee.

Change-Id: I9b1b1d678b639c47548752089f6a140f3dff0db0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e8d8b064e833f12a5a65c57a750bd8356fbe08a9
Original-Change-Id: If6c877df5a87df1b348de92868b91eed4a76de55
Original-Signed-off-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Reviewed-on: https://review.coreboot.org/23573
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Justin TerAvest <teravest@chromium.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/903762
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>