UPSTREAM: armv8: Allow for ARMv8 extensions to be chosen
There are now a few architectural extensions available for ARMv8, some
of which introduce instructions or other features that may be useful.
This allows the user to select an extension implemented on their SoC
which will set the -march option passed into the compiler.

BUG=none
BRANCH=none
TEST=none

Change-Id: I715d2a46f5944ebfc3c0639a04498a7a940a2fc5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: cc3b69bd14db6d975c35aebf93c3a3f1525c08fd
Original-Change-Id: Ifca50dad98aab130ac04df455bac2cfb65abf82e
Original-Signed-off-by: David Hendricks <dhendricks@fb.com>
Original-Reviewed-on: https://review.coreboot.org/23641
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/916191
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>